requirements.txt
langchain
langchain-groq
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
requests
beautifulsoup4
streamlit
python-dotenv
fpdf2